requirements.txt
kaggle
openai
python-dotenv
Pillow
matplotlib
sentence-transformers
transformers
flask
qdrant-client